Feat : 스터디룸 화상채팅 구현
(백엔드서버 연결 x)
