+++ src/pages/Room.jsx
@@ -4,7 +4,6 @@ import play from '../asset/play.svg';
 import camOn from '../asset/camon.svg';
 import camoff from '../asset/camoff.svg';
 import chat from '../asset/chat.svg';
-import Chevron from '../asset/Chevron.svg';
 import micOn from '../asset/micon.svg';
 import micoff from '../asset/micoff.svg';
 import setting from '../asset/setting2.svg';
@@ -14,33 +13,24 @@ import send from '../asset/send.svg';
 import cancel from '../asset/cancel.svg';
 import logo from '../asset/logo.svg';
 import profileimg from '../asset/user.svg';
-import viewmic from '../asset/viewmic.svg';
 import Vector from '../asset/Vector.svg';
-import { useParams } from 'react-router-dom';
-import { useQuery } from 'react-query';
-import { getRoomDetail } from '../api/api';
+import { useParams, useNavigate, useLocation } from 'react-router-dom';
 import UserVideoComponent from '../components/UserVideoComponent';
-function Room({
-  roomData,
-  publisher,
-  handleMainVideoStream,
-  mainStreamManager,
-  leaveSession,
-}) {
+import { useEffect, useRef } from 'react';
+import axios from 'axios';
+import { OpenVidu } from 'openvidu-browser';
+const APPLICATION_SERVER_URL =
+  process.env.NODE_ENV === 'production' ? '' : 'https://demos.openvidu.io/';
+function Room() {
   const [micon, setMicOn] = useState(false);
   const [camon, setCamOn] = useState(false);
   const [ischatOpen, setisChatOpen] = useState(false);
   const params = useParams();
-  //   console.log(params.id);
-  // const { isLoading, isError, data } = useQuery('room', () => getRoomDetail(params.id));
-  // if (isLoading) {
-  //   return <p>로딩중입니다....!</p>;
-  // }
+  const location = useLocation();
 
-  // if (isError) {
-  //   return <p>오류가 발생하였습니다...!</p>;
-  // }
-  // const roomData = data?.data;
+  const { roomData } = location.state;
+  console.log(roomData);
+
   const micbuttonhandler = () => {
     setMicOn(!micon);
   };
@@ -48,6 +38,194 @@ function Room({
   const cambuttonhandler = () => {
     setCamOn(!camon);
   };
+
+  const [state, setState] = useState({
+    mySessionId: 'SessionA',
+    myUserName: 'Participant' + Math.floor(Math.random() * 100),
+    session: undefined,
+    mainStreamManager: undefined,
+    publisher: undefined,
+    subscribers: [],
+  });
+  const navigate = useNavigate();
+
+  const OV = useRef(null);
+
+  const getUserName = localStorage.getItem('member');
+
+  useEffect(() => {
+    window.addEventListener('beforeunload', onbeforeunload);
+
+    return () => {
+      window.removeEventListener('beforeunload', onbeforeunload);
+    };
+  }, []);
+
+  const onbeforeunload = (event) => {
+    leaveSession();
+  };
+
+  const handleChangeSessionId = (e) => {
+    setState((prevState) => ({ ...prevState, mySessionId: e.target.value }));
+  };
+
+  const handleChangeUserName = (e) => {
+    setState((prevState) => ({ ...prevState, myUserName: e.target.value }));
+  };
+
+  const handleMainVideoStream = (stream) => {
+    if (state.mainStreamManager !== stream) {
+      setState((prevState) => ({ ...prevState, mainStreamManager: stream }));
+    }
+  };
+
+  const deleteSubscriber = (streamManager) => {
+    let subscribers = state.subscribers;
+    let index = subscribers.indexOf(streamManager, 0);
+    if (index > -1) {
+      subscribers.splice(index, 1);
+      setState((prevState) => ({ ...prevState, subscribers: subscribers }));
+    }
+  };
+
+  const joinSession = () => {
+    OV.current = new OpenVidu();
+    const mySession = OV.current.initSession();
+    setState((prevState) => ({ ...prevState, session: mySession }));
+  };
+
+  useEffect(() => {
+    joinSession();
+  }, []);
+
+  useEffect(() => {
+    if (state.session) {
+      const handleStream = (event) => {
+        let subscriber = state.session.subscribe(event.stream, undefined);
+        setState((prevState) => ({
+          ...prevState,
+          subscribers: [...prevState.subscribers, subscriber],
+        }));
+      };
+
+      const handleStreamDestroyed = (event) => {
+        deleteSubscriber(event.stream.streamManager);
+      };
+
+      const handleException = (exception) => {
+        console.warn(exception);
+      };
+
+      state.session.on('stream', handleStream);
+      state.session.on('streamDestroyed', handleStreamDestroyed);
+      state.session.on('exception', handleException);
+
+      (async function connectToken() {
+        try {
+          const token = await getToken();
+          await state.session.connect(token, { clientData: state.myUserName });
+
+          const publisher = await OV.current.initPublisherAsync(undefined, {
+            audioSource: undefined,
+            videoSource: undefined,
+            publishAudio: true,
+            publishVideo: true,
+            resolution: '960x540',
+            frameRate: 30,
+            insertMode: 'APPEND',
+            mirror: false,
+          });
+
+          state.session.publish(publisher);
+
+          const devices = await OV.current.getDevices();
+          const videoDevices = devices.filter((device) => device.kind === 'videoinput');
+          const currentVideoDeviceId = publisher.stream
+            .getMediaStream()
+            .getVideoTracks()[0]
+            .getSettings().deviceId;
+          const currentVideoDevice = videoDevices.find(
+            (device) => device.deviceId === currentVideoDeviceId
+          );
+
+          setState((prevState) => ({
+            ...prevState,
+            currentVideoDevice: currentVideoDevice,
+            mainStreamManager: publisher,
+            publisher: publisher,
+          }));
+        } catch (error) {
+          console.log(
+            '세션에 연결하는 중 오류가 발생했습니다:',
+            error.code,
+            error.message
+          );
+        }
+      })();
+
+      return () => {
+        state.session.off('stream', handleStream);
+        state.session.off('streamDestroyed', handleStreamDestroyed);
+        state.session.off('exception', handleException);
+      };
+    }
+  }, [state.session]);
+
+  const leaveSession = () => {
+    const mySession = state.session;
+
+    if (mySession) {
+      mySession.disconnect();
+    }
+    navigate(-1);
+    OV.current = null;
+    setState({
+      session: undefined,
+      subscribers: [],
+      mySessionId: 'SessionA',
+      myUserName: 'Participant' + Math.floor(Math.random() * 100),
+      mainStreamManager: undefined,
+      publisher: undefined,
+    });
+  };
+
+  const switchCamera = async () => {
+    try {
+      const devices = await OV.current.getDevices();
+      const videoDevices = devices.filter((device) => device.kind === 'videoinput');
+
+      if (videoDevices && videoDevices.length > 1) {
+        const newVideoDevice = videoDevices.filter(
+          (device) => device.deviceId !== state.currentVideoDevice.deviceId
+        );
+
+        if (newVideoDevice.length > 0) {
+          const newPublisher = OV.current.initPublisher(undefined, {
+            videoSource: newVideoDevice[0].deviceId,
+            publishAudio: true,
+            publishVideo: true,
+            mirror: true,
+          });
+
+          await state.session.unpublish(state.mainStreamManager);
+
+          await state.session.publish(newPublisher);
+          setState((prevState) => ({
+            ...prevState,
+            currentVideoDevice: newVideoDevice[0],
+            mainStreamManager: newPublisher,
+            publisher: newPublisher,
+          }));
+        }
+      }
+    } catch (e) {
+      console.error(e);
+    }
+  };
+
+  const { mySessionId, myUserName, mainStreamManager, publisher, subscribers, session } =
+    state;
+
   return (
     <Stcontainer>
       <StLayout>
@@ -67,22 +245,15 @@ function Room({
           </Stheader>
           <Stcamarea ischatOpen={ischatOpen}>
             {/* <Stcambox> */}
-            <Stcamboxname>닉네임</Stcamboxname>
-            {mainStreamManager !== undefined ? (
-              <div id="main-video" className="col-md-6">
-                <UserVideoComponent streamManager={mainStreamManager} />
-              </div>
-            ) : null}
-            <img src={viewmic} alt="" width={36} height={36} />
-            {/* </Stcambox> */}
             {publisher !== undefined ? (
               <div
                 className="stream-container col-md-6 col-xs-6"
                 onClick={() => handleMainVideoStream(publisher)}
               >
-                <UserVideoComponent streamManager={publisher} />
+                <UserVideoComponent streamManager={publisher} userName={getUserName} />
               </div>
             ) : null}
+            {/* </Stcambox> */}
           </Stcamarea>
           <Stfooter>
             <Stsettingbox>
@@ -162,6 +333,48 @@ function Room({
       </StLayout>
     </Stcontainer>
   );
+  async function getToken() {
+    try {
+      const sessionId = await createSession(mySessionId);
+      const response = await createToken(sessionId);
+      console.log('4' + response);
+      return response;
+    } catch (error) {
+      console.error('인터넷 요청이 실패했습니다: getToken');
+    }
+  }
+
+  async function createSession(sessionId) {
+    try {
+      const response = await axios.post(
+        APPLICATION_SERVER_URL + 'api/sessions',
+        { customSessionId: sessionId },
+        {
+          headers: { 'Content-Type': 'application/json' },
+        }
+      );
+      console.log('2' + sessionId);
+      return response.data; // The sessionId
+    } catch (error) {
+      console.error('인터넷 요청이 실패했습니다: createSession');
+    }
+  }
+
+  async function createToken(sessionId) {
+    try {
+      const response = await axios.post(
+        APPLICATION_SERVER_URL + 'api/sessions/' + sessionId + '/connections',
+        {},
+        {
+          headers: { 'Content-Type': 'application/json' },
+        }
+      );
+      console.log('3' + sessionId);
+      return response.data; // The token
+    } catch (error) {
+      console.error('인터넷 요청이 실패했습니다: createToken');
+    }
+  }
 }
 
 export default Room;
